+++ main.py
@@ -3,4 +3,4 @@ from panda3d.core import loadPrcFile
 loadPrcFile("VisorConfig.prc")
 
 # Run main program
-import src.program
+import src.visorview
